Load template files relative to exe location
Also adds the .xml extension when the template file passed in does not
have an extension. This is so we can add the template files in the same
folder as the MGPG executable (in a subfolder of the MG tools folder) and
people that add the MG tools folder to their PATH can just call e.g.
`MGPG Windows MyGame/` to generate a windows project, regardless of the
working directory.